+++ .env.development
@@ -4,8 +4,13 @@ DB_USERNAME=gimptrade
 DB_PASSWORD=wX8YBskCUT
 DB_DATABASE=gimp-trade
 
+BITMEX_API_URL=https://www.bitmex.com/api/v1
 BITMEX_API_KEY=
 BITMEX_API_SECRET=
+
+UPBIT_API_URL=https://crix-api-endpoint.upbit.com/v1
+
+FREEFORE_API_URL=https://www.freeforexapi.com/api
 
 FIXED_USDKRW=1150
