+++ package.json
@@ -33,10 +33,10 @@
     "@nestjs/schedule": "^0.4.0",
     "@nestjs/serve-static": "^2.1.1",
     "@nestjs/typeorm": "^7.1.0",
+    "@types/dotenv": "^8.2.0",
     "@types/qs": "^6.9.3",
     "axios": "^0.19.2",
     "crypto": "^1.0.1",
-    "@types/dotenv": "^8.2.0",
     "dotenv": "^8.2.0",
     "moment": "^2.26.0",
     "pg": "^8.2.1",
