+++ src/utils/math/gimp.ts
@@ -1,0 +1,3 @@
+export default function getGimp(btcKrw: number, usdKrw: number, btcUsd:number): number {
+  return Number((((btcKrw - usdKrw * btcUsd) * 100) / btcKrw).toFixed(3))
+}
